// generator/lib/src/model/ids/state.ts
export type StateIdentifier =
  | "STATE_NORMAL"
  | "STATE_SLEEPING"
  | "STATE_BERSERK"
  | "STATE_PANIC"
  | "STATE_STUNNED"
  | "STATE_INVISIBLE"
  | "STATE_HELPLESS"
  | "STATE_IMMOBILE"
  | "STATE_FROZEN_DEATH"
  | "STATE_STONE_DEATH"
  | "STATE_EXPLODING_DEATH"
  | "STATE_FLAME_DEATH"
  | "STATE_ACID_DEATH"
  | "STATE_DEAD"
  | "STATE_REALLY_DEAD"
  | "STATE_SILENCED"
  | "STATE_CHARMED"
  | "STATE_POISONED"
  | "STATE_HASTED"
  | "STATE_SLOWED"
  | "STATE_INFRAVISION"
  | "STATE_BLIND"
  | "STATE_DISEASED"
  | "STATE_FEEBLEMINDED"
  | "STATE_HARMLESS"
  | "STATE_NONDETECTION"
  | "STATE_IMPROVEDINVISIBILITY"
  | "STATE_NOT_TARGETABLE"
  | "STATE_BLESS"
  | "STATE_CHANT"
  | "STATE_DRAWUPONHOLYMIGHT"
  | "STATE_LUCK"
  | "STATE_AID"
  | "STATE_CHANTBAD"
  | "STATE_BLUR"
  | "STATE_MIRRORIMAGE"
  | "STATE_ILLUSIONS"
  | "STATE_CONFUSED"
  | "STATE_RANGED_TARGET"
  | "CD_STATE_NOTVALID"
  | "STATE_DISABLED"
  | "STATE_DISABLED_CASTING"
  | "STATE_DEBUFF"
  | "STATE_NOT_APPROACHABLE";